+++ NEWS.md
@@ -1,5 +1,6 @@
 # sigminer 2.0.6
 
+- Added `sig_unify_extract()` as an unified signature extractor.
 - Fixed error showing reference signature profile for `CNS_TCGA` database.
 
 # sigminer 2.0.5